+++ packages/plugin-rax-app/src/getPostCssPlugin.js
@@ -18,7 +18,7 @@ module.exports = function getPlugins(type) {
     case 'web':
       return [
         atImport(),
-        require('postcss-preset-env')({
+        require('@builder/pack/deps/postcss-preset-env')({
           autoprefixer: {
             flexbox: 'no-2009',
           },