+++ packages/plugin-rax-web/src/setEntry.ts
@@ -9,6 +9,11 @@ export default (config, context) => {
   if (!web.mpa) {
     // SPA
     const { entryPath } = getAppEntry(rootDir);
+
+    if (!entryPath) {
+      throw new Error('项目目录缺少 src/app.ts 入口文件!');
+    }
+
     const entryConfig = config.entry('index');
 
     entryConfig.add(entryPath);