+++ packages/plugin-ssr/src/ssr/EntryLoader.ts
@@ -22,18 +22,38 @@ function addDefineInitialPage() {
   let Page;
   if (enableRouter) {
     const route = staticConfig.routes.find(({ path }) => path === pathname);
-    Page = route.component;
+    Page = route.component();
   } else {
     Page = appConfig.app && appConfig.app.renderComponent;
   }
   `;
 }
 
+function addImportSource(source, exportPageComponent: boolean): string {
+  if (exportPageComponent) {
+    return `import PageComponent from '${formatPath(source)}';`;
+  }
+
+  return `import '${formatPath(source)}';`;
+}
+
+function addExecuteRunApp(exportPageComponent: boolean): string {
+  if (exportPageComponent) {
+    return `runApp({
+      app: {
+        renderComponent: PageComponent,
+      }
+    });`;
+  }
+
+  return '';
+}
+
 /**
  * Global variables:
  * utils: Generator, parseUrl
  * rax render: createElement/renderer
- * Component: Page/Document
+ * Component: Document/PageComponent
  * generated in .rax: appConfig/staticConfig/createBaseApp/emitLifeCycles/setHistory/enableRouter
  * MPA: pageConfig
  */
@@ -50,8 +70,10 @@ export default function () {
     import { getAppConfig } from '${formatPath(path.join(corePath, 'appConfig'))}';
     import { emitLifeCycles } from '${formatPath(path.join(corePath, 'publicAPI'))}';
     import { setHistory } from '${formatPath(path.join(corePath, 'routerAPI'))}';
-    import '${formatPath(this.resourcePath)}';
-    import app from '${formatPath(formattedQuery.runAppPath)}';
+    ${addImportSource(this.resourcePath, formattedQuery.exportPageComponent)}
+    import app, { runApp } from '${formatPath(formattedQuery.runAppPath)}';
+
+    ${addExecuteRunApp(formattedQuery.exportPageComponent)}
 
     const { createBaseApp, staticConfig, pageConfig, TabBar, enableRouter } = app;
 
@@ -81,6 +103,12 @@ export default function () {
 
     function getInitialContext(ctx) {
       const { req, res } = ctx;
+
+      // 某些调用环境 mock 的 request 上不存在 url 字段
+      if (!req.url) {
+        return ctx;
+      }
+
       const { hash, search, query } = parseUrl(req.url);
       const pathname = req.path;
       const location = {
@@ -91,27 +119,27 @@ export default function () {
       };
 
       return {
-        req,
-        res,
         pathname,
         query,
         location,
+        ...ctx,
       };
     }
 
     ${formattedQuery.documentPath ? addCustomRenderComponentToHTML(formattedQuery) : addBuiltInRenderComponentToHTML(formattedQuery)}
 
     async function renderToHTML(req, res, options = {}) {
-      const { initialData, htmlTemplate, chunkInfo } = options;
+      const { initialData, htmlTemplate } = options;
       ${addDefineInitialPage()}
-      const html = await renderComponentToHTML(Page, getInitialContext({ req, res }), initialData, htmlTemplate, chunkInfo);
+      const html = await renderComponentToHTML(Page, getInitialContext({ req, res }), initialData);
       return html;
     }
 
     async function renderToHTMLWithContext(ctx, options = {}) {
-      const { initialData, htmlTemplate, chunkInfo } = options;
+      const { req, res } = ctx;
+      const { initialData, htmlTemplate } = options;
       ${addDefineInitialPage()}
-      const html = await renderComponentToHTML(Page, getInitialContext(ctx), initialData, htmlTemplate, chunkInfo);
+      const html = await renderComponentToHTML(Page, getInitialContext(ctx), initialData, htmlTemplate);
       return html;
     }
 
@@ -130,12 +158,12 @@ export default function () {
       ctx.body = html;
     }
 
-    async function render(ctx, options) {
+    async function render(ctx, options = {}) {
       let html;
       if (ctx.req) {
-        const { initialData, htmlTemplate, chunkInfo } = options;
+        const { initialData, htmlTemplate } = options;
         try {
-          html = await renderToHTML(ctx.req, ctx.res, { initialData, htmlTemplate, chunkInfo });
+          html = await renderToHTML(ctx.req, ctx.res, { initialData, htmlTemplate });
         } catch (e) {
           html = htmlTemplate;
           console.error(e);