+++ packages/plugin-ssr/src/ssr/addCustomRenderComponentToHTML.ts
@@ -13,13 +13,13 @@ export default function addCustomRenderComponentToHTML(
     updateDataInClient,
   }: IFormattedLoaderQuery,
 ) {
-  const injectedScripts = [];
+  const injectedScripts = (injectedHTML.scripts || []);
 
   if (injectedHTML.comboScripts) {
     injectedScripts.unshift(genComboedScript(injectedHTML.comboScripts));
   }
   return `
-  async function renderComponentToHTML(Component, ctx, initialData, htmlTemplate, chunkInfo = {}) {
+  async function renderComponentToHTML(Component, ctx, initialData, htmlTemplate) {
     const pageInitialProps = await getInitialProps(Component, ctx);
     const data = {
       __SSR_ENABLED__: true,
@@ -37,15 +37,12 @@ export default function addCustomRenderComponentToHTML(
     }
     const title = ${JSON.stringify(pageConfig.window?.title || '')} || getTitle(staticConfig);
 
-    if (process.env.NODE_ENV !== 'development') {
-      // chunk info will be replaced during local build
-      chunkInfo = JSON.parse(decodeURIComponent("__CHUNK_INFO__"));
-    }
+    const chunkInfo = JSON.parse(decodeURIComponent("__CHUNK_INFO__"));
 
     let scripts = chunkInfo[${JSON.stringify(entryName)}].js
       .map(filename => ${JSON.stringify(publicPath)} + filename);
     let styles = chunkInfo[${JSON.stringify(entryName)}].css
-      .map(filename => ${JSON.stringify(publicPath)} + filename);;
+      .map(filename => ${JSON.stringify(publicPath)} + filename);
 
     ${assetsProcessor}