+++ CHANGELOG.md
@@ -1,5 +1,81 @@
 ## Changelog
 
+## 3.8.7
+
+- Fix: SPA SSR `getInitialProps` is invalid
+- Feat: miniapp webview mode should modify app to MPA type
+- Chore: add ssr test case
+- Chore: remove DEF comment node
+
+## 3.8.6
+
+- Fix: page component import method in the env other than web
+- Fix: MPA runApp file path in SSR
+- Fix: override developer custom title
+
+## 3.8.5
+
+- Fix: special SPA route.path
+
+## 3.8.4
+
+- Chore: update `parse-url` to v6
+- SSR: fix inject configured scripts
+
+## 3.8.3
+
+- Fix: remove comment node in html result
+
+## 3.8.2
+
+- Fix: ssr context shouldn't be overrode
+- Fix: static export mode build error
+
+## 3.8.1
+
+SSR: fix chunkInfo is undefined FaaS app
+SSR: fix request might be fake that lose url field
+SSR: fix devServer.hot should be closed
+SSR: fix appConfig might be undefined
+Fix: __webpack_public_path__ is invalid in front of page file
+Fix: tabBar show in other page
+Fix: child class need use declare with add type for property which extends parent
+Fix: miniapp compile mode @alias is invalid
+Fix: invalid host in bind host case
+Chore: use webpack-chain directly
+
+## 3.8.0
+
+- Feat: support toggle webpack version 5 by `webpack5: true`
+- Feat: support use swc instead of babel by `swc: true`
+- Feat: `dropLogLevel` for removing different level log code
+- Feat: miniapp support webview mode
+- Feat: support MPA runtime plugin
+- Feat: SPA supports per-page configurations of `keepAlive` and `lazy`, such as
+  ```json
+  // app.json
+  {
+    "routes": [
+      {
+        "path": "/",
+        "source": "pages/Home/index",
+        "keepAlive": true,
+        "lazy": false,
+      }
+    ]
+  }
+  ```
+- Chore: MPA changed to render every page by `runApp`
+- Chore: use `minify` `object` type options instead `terserOptions`, such as
+  ```json
+  // build.json
+  "minify": {
+    "type": "terser",
+    "options": {}
+  }
+  ```
+- Chore: only wrapper DOM/BOM fake API in specified page bundle with build miniapp
+
 ## 3.7.4
 
 - Chore: add node version validate