+++ packages/plugin-rax-web/CHANGELOG.md
@@ -1,5 +1,27 @@
 # Changelog
 
+## v2.0.4
+
+- Fix: not filter routes by target in document plugin
+- Fix: DEF remove comment node check
+
+## v2.0.3
+
+- Chore: add logic protect
+
+## v2.0.2
+
+- Fix: remove comment node
+
+## v2.0.1
+
+- Fix: static export error
+
+## v2.0.0
+
+- Feat: upgrade dependencies for webpack 5
+- Feat: use `runApp` as MPA default render trigger
+
 ## v1.4.8
 
 - Fix: `publicPath` in document